src/requirements.txt
pysftp
python-gnupg
